requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.36
pyarrow>=12.0
statsmodels>=0.14
pmdarima>=2.0
arch>=6.3
hmmlearn>=0.3
tensorflow>=2.13
scikit-learn>=1.3
ta>=0.10
antropy>=0.1.6
plotly>=5.15
dash>=2.14
dash-bootstrap-components>=1.5
flask-caching>=2.1
joblib>=1.3
pyyaml>=6.0
scipy>=1.11
pytest>=7.4
